\name{print.regress}
\alias{print.regress}

\title{Print a summary of regress output}

\description{Summarises output from \code{regress}.}

\synopsis{print.regress(x, digits=3, fixed.effects=T, ...)}

\usage{print.regress(x,digits=4,fixed.effects=F, ...)}

\arguments{
  \item{x}{Output from \code{regress}.}
  \item{digits}{Number of significant decimal places to print.}
  \item{fixed.effects}{Logical, include the summary of fixed effects
    also.}
  \item{...}{additional arguments to be passed}
}

\author{David Clifford}

\keyword{regression}
\keyword{models}
\keyword{multivariate}

