\name{summary.regress}
\alias{summary.regress}

\title{Summarise a Gaussian Linear Model from regress}

\description{Summarises output from \code{regress}.}

\synopsis{summary.regress(object, ...)}

\usage{summary.regress(object, ...)}

\arguments{
  \item{object}{Output from \code{regress}.}
  \item{...}{additional parameters as needed}
}

\details{ All summary information included in the \code{regress}
  function itself.  This function is included for completeness.  }

\author{David Clifford}

\keyword{regression}
\keyword{models}
\keyword{multivariate}

