\name{summary.regress}
\alias{summary.regress}

\title{Summarise a Gaussian Linear Model from regress}

\description{Summarises output from \code{regress}.}

\synopsis{summary.regress(object, digits=3, fixed.effects=T,...)}

\usage{summary.regress(object,digits=4,fixed.effects=F)}

\arguments{
  \item{object}{Output from \code{regress}.}
  \item{digits}{Number of significant decimal places to print.}
  \item{fixed.effects}{Logical, include the summary of fixed effects also.}
}

\author{David Clifford}

\keyword{regression}
\keyword{models}
\keyword{multivariate}

